fe: split GlobalDataContext, add useGlobalData(), clean up console.log
Three small things that had been sitting there the whole session:

- GlobalDataContext.tsx exported both the Context object and the
  GlobalDataProvider component from the same file, which is exactly
  what triggers eslint's react-refresh/only-export-components warning
  - it's shown up in every single `npm run lint` run since ESLint was
  added. Split into globalDataContext.ts (just the Context + type) and
  GlobalDataProvider.tsx (just the component, renamed to match what it
  actually exports now - also sidesteps a real TS error, "differs only
  in casing", from having globalDataContext.ts and GlobalDataContext.tsx
  coexist).
- Added useGlobalData() (useGlobalData.ts): every consumer
  (Navigation, Login, Signup) was repeating the exact same
  useContext(GlobalDataContext) + "throw if undefined" boilerplate.
  Centralized once instead of copy-pasted three times.
- Replaced console.log('error')/console.log(err) with console.error
  and an actual message - these were being used for real error/failure
  paths (a 500 response, a caught fetch rejection), just logged with
  the wrong console method and no context.

Verified: typecheck, lint (0 warnings now - the react-refresh warning
that's persisted all session is gone), prettier, and build all pass.
Drove a full signup -> home -> logout flow in a real browser against
the actual built frontend afterward to confirm no regression: session
worked, nav rendered correctly, logout redirected properly, zero
console errors.
